fix add Item PO

+++ src/main/java/sistem/operasional/sioperasional/controller/PurchaseOrderController.java
@@ -108,7 +108,7 @@ public class PurchaseOrderController {
                            @PathVariable("nomorPurchaseOrder") String nomorPO,
                            @ModelAttribute PurchaseOrderModel purchaseOrder,
                            Model model) {
-
+        logger.info("=========================MAU MASUK LOOP 1");
         HashMap<String, ItemPOModel> itemPOMap = new HashMap<>();
         PurchaseOrderModel purchaseOrder1 = new PurchaseOrderModel();
         for (PurchaseOrderModel po : purchaseOrderService.getAll()) {
@@ -116,12 +116,15 @@ public class PurchaseOrderController {
                 purchaseOrder1 = po;
             }
         }
-        List<ItemPOModel> oldItemPOModel = itemPOService.getAllItemPO();
-        for (ItemPOModel item : oldItemPOModel) {
-            if(item.getPurchaseOrder().getNomorPurchaseOrder().equalsIgnoreCase(purchaseOrder1.getNomorPurchaseOrder()))
-            itemPODB.delete(item);
+        logger.info("=========================MASUK SETELAH LOOP 1");
+        if(itemPOService.getAllItemPO() != null) {
+            List<ItemPOModel> oldItemPOModel = itemPOService.getAllItemPO();
+            for (ItemPOModel item : oldItemPOModel) {
+                if (item.getPurchaseOrder().getNomorPurchaseOrder().equalsIgnoreCase(purchaseOrder1.getNomorPurchaseOrder()))
+                    itemPODB.delete(item);
+            }
         }
-
+        logger.info("=========================MASUK SETELAH IF ");
         for (ItemPOModel itemPO : purchaseOrder.getListItemPO()) {
             itemPO.setPurchaseOrder(purchaseOrder1);
             String kategoriItem = itemPO.getKategoriItem().getNamaKategoriItem();
@@ -129,7 +132,7 @@ public class PurchaseOrderController {
 //            temp = itemPO;
             itemPOService.addItemPO(itemPO);
         }
-
+        logger.info("=========================MASUK SETELAH LOOP 2");
         String namaVendor = purchaseOrder1.getVendor().getNamaVendor();
         List<ItemPOModel> listItemPO = purchaseOrder1.getListItemPO();
